+++ src/main/java/com/example/examensarbete/entities/Ingredient.java
@@ -18,13 +18,13 @@ import java.util.Objects;
 public class Ingredient implements Serializable {
     @Id
     @Column(name = "ingredient_id")
-    private Long id;
+    private Integer id;
 
     @NotEmpty
     @Column(name = "ingredient_name")
     private String name;
 
-    public Ingredient(Long id, String name) {
+    public Ingredient(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
